Rename bundle to charm archive in the uniter.
Within the uniter, many files refer to a bundle. However the uniter
doesn't deploy bundles. In this case, it's an old name to what charm
archives used to be called in 2014. Before a bundle exists I'd guess.

The old naming is confusing, change it over to the more modern naming:
  Bundle => CharmArchive
  BundleInfo => CharmInfo
  BundleReader => CharmReader