Resolve charmstore to charmhub changes with bundle diff tests.
Charmstore charms act different from charmhub charms so that a direct
replacement doesn't work especially when comparing against model data of
a deployed charm. Bundles must specify revisions and channels for
charmhub charms when compared against an existing deployment.

There looks to be a bug here, as "ch:charm" is the same as "charm" and
yet they produce a difference where there is none.