fix: wire up EnvironConfigGetter to use model config service
This is a stop gap on the way to reducing dependence on Environs and
replacing with appropriate calls into the different domain services
where possible. That work was started in PR 18013.

GetNewPolicyFunc needs a model config service getter. Callers do not
have a model at the setup time of GetNewPolicyFunc. Use a model
config service getter instead.

Special case the model config service getter with the environStatePolicy
structure to limit use and not require everywhere, using a model config
service is preferable.

A short term solution one the way to removing the Policy's all together.

Delete stateenvirons NewInstancePrechecker and associated code. The
functionality has been replace and no longer required here.

Remove configvalidator_test.go. These are model config tests which are
failing in state after changes to the EnvironConfigGetter. We're near
the end of the work to move model config from state to a domain. Thus
they can be removed now rather than fixing.