feat: add juju-info to every charm
Only if the Requires Relation is not defined in the charm metadata.
Previously juj-info was an implicit relation that was not put into state
and added in to the allowed list when verifying a relation was possible.
Now that we have referential integrity, it is necessary to add for each
new charm.