feat: implement ModifyOfferAccess using domains only
Implemented in the facade as the permission data, both for allowing and
persisting the update is in the controller database. While each offer
must be found in its specific model database.

Note: there is no referential integrity on offer permissions. They are
stored in the controller database with users and permission data, which
the offers are persistented in the model databases.